createButtonPane2() in AbstractOkCancelDialog with the Help button; 'Help' in some plug-ins with dialogs.
