+++ plugin/src/org/aavso/tools/vstar/external/plugin/GAIADR2XformObSource.java
@@ -46,6 +46,7 @@ import org.aavso.tools.vstar.ui.dialog.AbstractOkCancelDialog;
 import org.aavso.tools.vstar.ui.dialog.MessageBox;
 import org.aavso.tools.vstar.ui.dialog.TextArea;
 import org.aavso.tools.vstar.ui.dialog.TextField;
+import org.aavso.tools.vstar.ui.mediator.Mediator;
 import org.aavso.tools.vstar.util.Pair;
 
 public class GAIADR2XformObSource extends GaiaObSourceBase {
@@ -172,8 +173,8 @@ public class GAIADR2XformObSource extends GaiaObSourceBase {
 
 			topPane.add(createAdditiveLoadCheckboxPane(additiveChecked));
 
-			// OK, Cancel
-			topPane.add(createButtonPane());
+			// Help, Cancel, OK
+			topPane.add(createButtonPane2());
 
 			contentPane.add(topPane);
 
@@ -324,7 +325,15 @@ public class GAIADR2XformObSource extends GaiaObSourceBase {
 			
 			super.showDialog();			
 		}
-		
+
+		/**
+		 * @see org.aavso.tools.vstar.ui.dialog.AbstractOkCancelDialog#helpAction()
+		 */
+		@Override
+		protected void helpAction() {
+			Mediator.openPluginHelp(getDocName());
+		}
+
 		/**
 		 * @see org.aavso.tools.vstar.ui.dialog.AbstractOkCancelDialog#cancelAction()
 		 */
